+++ forgero-fabric-compat/src/main/java/com/sigmundgranaas/forgero/fabric/patchouli/GuideBookGenerator.java
@@ -1,15 +1,22 @@
-package com.sigmundgranaas.forgero.minecraft.common.recipe.implementation.generator;
+package com.sigmundgranaas.forgero.fabric.patchouli;
 
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 import com.sigmundgranaas.forgero.core.Forgero;
+import com.sigmundgranaas.forgero.minecraft.common.recipe.RecipeCreator;
 import com.sigmundgranaas.forgero.minecraft.common.recipe.RecipeGenerator;
 import com.sigmundgranaas.forgero.minecraft.common.recipe.RecipeWrapper;
 import com.sigmundgranaas.forgero.minecraft.common.recipe.customrecipe.RecipeTypes;
+import com.sigmundgranaas.forgero.minecraft.common.recipe.implementation.generator.TemplateGenerator;
 import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.util.Identifier;
 
+import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+
 public class GuideBookGenerator implements RecipeGenerator {
+
     private final TemplateGenerator generator;
     private final String value;
 
@@ -18,13 +25,20 @@ public class GuideBookGenerator implements RecipeGenerator {
         this.value = value;
     }
 
+    public static void registerGuideBookRecipes() {
+        List<RecipeGenerator> recipes = Stream.of("binding", "pickaxe_head", "shovel_head", "axe_head", "hoe_head", "sword_blade", "handle")
+                .map(tag -> new GuideBookGenerator(RecipeCreator.INSTANCE.templates(), tag))
+                .collect(Collectors.toList());
+        RecipeCreator.INSTANCE.registerGenerator(recipes);
+    }
+
     @Override
     public RecipeWrapper generate() {
         JsonObject template = generator.generate(RecipeTypes.TOOLPART_SCHEMATIC_RECIPE).orElse(new JsonObject());
         template.addProperty("type", "patchouli:shapeless_book_recipe");
         JsonArray ingredients = template.getAsJsonArray("ingredients");
         JsonObject tag = new JsonObject();
-        tag.addProperty("tag", value);
+        tag.addProperty("tag", "forgero:" + value);
         JsonObject book = new JsonObject();
         book.addProperty("item", "minecraft:book");
 
@@ -32,11 +46,9 @@ public class GuideBookGenerator implements RecipeGenerator {
         ingredients.add(tag);
         template.add("ingredients", ingredients);
 
+
         template.addProperty("book", "forgero:forgero_guide");
-
-        Identifier id = new Identifier(Forgero.NAMESPACE, "forgero_guide_book_recipe_" + value);
-
-        return RecipeWrapper.of(id, tag, RecipeTypes.MISC_SHAPELESS);
+        return RecipeWrapper.of(new Identifier(Forgero.NAMESPACE, "forgero_guide_book_recipe_" + value), template, RecipeTypes.MISC_SHAPELESS);
     }
 
     @Override